MovieVerse-Middleware/requirements.txt
Django~=5.0.6
djangorestframework==3.12
gunicorn==20.1.0
psycopg2-binary==2.8.6
pytz==2021.1
flask~=3.0.3
redis==5.0.4
python-jose==3.3.0
pydantic==2.7.1
pydantic-settings==2.2.1
starlette==0.37.2
